\name{eglen2015-package}
\alias{eglen2015-package}
\alias{eglen2015}
\docType{package}
\title{
Data and code for the Eglen (2015) review
}
\description{
This package contains the code, data and manuscript for the Eglen (2015)
review article.
}
\details{
\tabular{ll}{
Package: \tab eglen2015\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-09\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Stephen Eglen

Maintainer: Stephen Eglen <sje30@cam.ac.uk>
}
\references{
Any extra references here.
}
\keyword{ package }
\examples{
}
